requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# MongoDB driver (Motor deprecated in 2026, using PyMongo instead)
pymongo==4.6.3
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
redis==5.0.1
celery==5.3.4
openai==1.3.5
pillow==10.1.0
opencv-python==4.8.1.78
httpx==0.25.2
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
# email-validator==2.1.0.post1  # Now included with pydantic[email]
pyzbar==0.1.9
certifi==2024.2.2
dnspython==2.4.2
firebase-admin==6.1.0
slowapi==0.1.9
# zeptomail==0.2.3  # Removed - using direct HTTP requests instead

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0